+++ packages/dart_mappable/test/basic_json/basic_json_test.mapper.g.dart
@@ -98,7 +98,7 @@ class _CarCopyWithImpl<$R> extends BaseCopyWith<Car, $R> implements CarCopyWith<
 class BrandMapper extends EnumMapper<Brand> {
   BrandMapper._();
 
-  @override  Brand fromString(String value) {
+  @override  Brand decode(dynamic value) {
     switch (value) {
       case 'toyota': return Brand.Toyota;
       case 'audi': return Brand.Audi;
@@ -107,7 +107,7 @@ class BrandMapper extends EnumMapper<Brand> {
     }
   }
 
-  @override  String toStringValue(Brand value) {
+  @override  dynamic encode(Brand value) {
     switch (value) {
       case Brand.Toyota: return 'toyota';
       case Brand.Audi: return 'audi';
@@ -117,6 +117,8 @@ class BrandMapper extends EnumMapper<Brand> {
 }
 
 extension BrandMapperExtension on Brand {
+  dynamic toValue() => Mapper.toValue(this);
+  @Deprecated('Use \'toValue\' instead')
   String toStringValue() => Mapper.toValue(this) as String;
 }