+++ packages/dart_mappable/test/enums/enums_test.dart
@@ -12,6 +12,18 @@ enum State { On, off, itsCOMPLICATED }
 @MappableEnum(caseStyle: CaseStyle.paramCase)
 enum Color { Green, BLUE, bloodRED }
 
+@MappableEnum(mode: ValuesMode.indexed)
+enum Items { first, second, third }
+
+@MappableEnum(mode: ValuesMode.indexed)
+enum Status {
+  zero,
+  @MappableValue(200)
+  success,
+  @MappableValue('error')
+  error
+}
+
 void main() {
   group('Enum Mappers', () {
     test('Should encode enums', () {
@@ -25,14 +37,6 @@ void main() {
       expect(Mapper.fromValue<State>('on'), equals(State.On));
       expect(Mapper.fromValue<State>('off'), equals(State.off));
       expect(Mapper.fromValue<State>('none'), equals(State.off));
-      expect(
-        () => Mapper.fromValue<State>(1),
-        throwsMapperException(MapperException.chain(
-          MapperMethod.decode,
-          '(State)',
-          MapperException.unexpectedType(int, State, 'String'),
-        )),
-      );
 
       expect(Mapper.fromValue<Color>('green'), equals(Color.Green));
       expect(Mapper.fromValue<Color>('blood-red'), equals(Color.bloodRED));
@@ -44,6 +48,17 @@ void main() {
           MapperException.unknownEnumValue('pink'),
         )),
       );
+    });
+
+    test('Should use index mode for enum', () {
+      expect(Items.second.toValue(), equals(1));
+      expect(Mapper.fromValue<Items>(2), equals(Items.third));
+    });
+
+    test('Should use custom values for enum', () {
+      expect(Status.zero.toValue(), equals(0));
+      expect(Mapper.toValue(Status.success), equals(200));
+      expect(Mapper.fromValue<Status>('error'), equals(Status.error));
     });
   });
 }