+++ packages/dart_mappable/test/enums/enums_test.mapper.g.dart
@@ -11,6 +11,8 @@ var _mappers = <BaseMapper>{
   // enum mappers
   StateMapper._(),
   ColorMapper._(),
+  ItemsMapper._(),
+  StatusMapper._(),
   // custom mappers
 };
 
@@ -25,7 +27,7 @@ var _mappers = <BaseMapper>{
 class StateMapper extends EnumMapper<State> {
   StateMapper._();
 
-  @override  State fromString(String value) {
+  @override  State decode(dynamic value) {
     switch (value) {
       case 'on': return State.On;
       case 'off': return State.off;
@@ -34,7 +36,7 @@ class StateMapper extends EnumMapper<State> {
     }
   }
 
-  @override  String toStringValue(State value) {
+  @override  dynamic encode(State value) {
     switch (value) {
       case State.On: return 'on';
       case State.off: return 'off';
@@ -44,13 +46,15 @@ class StateMapper extends EnumMapper<State> {
 }
 
 extension StateMapperExtension on State {
+  dynamic toValue() => Mapper.toValue(this);
+  @Deprecated('Use \'toValue\' instead')
   String toStringValue() => Mapper.toValue(this) as String;
 }
 
 class ColorMapper extends EnumMapper<Color> {
   ColorMapper._();
 
-  @override  Color fromString(String value) {
+  @override  Color decode(dynamic value) {
     switch (value) {
       case 'green': return Color.Green;
       case 'blue': return Color.BLUE;
@@ -59,7 +63,7 @@ class ColorMapper extends EnumMapper<Color> {
     }
   }
 
-  @override  String toStringValue(Color value) {
+  @override  dynamic encode(Color value) {
     switch (value) {
       case Color.Green: return 'green';
       case Color.BLUE: return 'blue';
@@ -69,7 +73,59 @@ class ColorMapper extends EnumMapper<Color> {
 }
 
 extension ColorMapperExtension on Color {
+  dynamic toValue() => Mapper.toValue(this);
+  @Deprecated('Use \'toValue\' instead')
   String toStringValue() => Mapper.toValue(this) as String;
+}
+
+class ItemsMapper extends EnumMapper<Items> {
+  ItemsMapper._();
+
+  @override  Items decode(dynamic value) {
+    switch (value) {
+      case 0: return Items.first;
+      case 1: return Items.second;
+      case 2: return Items.third;
+      default: throw MapperException.unknownEnumValue(value);
+    }
+  }
+
+  @override  dynamic encode(Items value) {
+    switch (value) {
+      case Items.first: return 0;
+      case Items.second: return 1;
+      case Items.third: return 2;
+    }
+  }
+}
+
+extension ItemsMapperExtension on Items {
+  dynamic toValue() => Mapper.toValue(this);
+}
+
+class StatusMapper extends EnumMapper<Status> {
+  StatusMapper._();
+
+  @override  Status decode(dynamic value) {
+    switch (value) {
+      case 0: return Status.zero;
+      case 200: return Status.success;
+      case 'error': return Status.error;
+      default: throw MapperException.unknownEnumValue(value);
+    }
+  }
+
+  @override  dynamic encode(Status value) {
+    switch (value) {
+      case Status.zero: return 0;
+      case Status.success: return 200;
+      case Status.error: return 'error';
+    }
+  }
+}
+
+extension StatusMapperExtension on Status {
+  dynamic toValue() => Mapper.toValue(this);
 }