+++ packages/dart_mappable/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.1.0
+
+- Added support for custom enum values
+  - Choose between `ValuesMode.named` and `ValuesMode.indexed` for automatic generation of encoded values
+  - Use `@MappableValue(myCustomValue)` on an enum value to specify a custom encoded value
+  - Deprecated the `String toStringValue()` extension method in favor of the more general `dynamic toValue()`
+
 # 1.0.3
 
 - Fixed issue with minified standard types on web